+++ src/main/java/com/example/apiweb/Neo4J/modelNeo/TutorNeoModel.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class TutorNeoModel {
 
     @Id
-    private Long id_tutor;
+    private Integer id_tutor;
     private String nombre_tutor;
 }
