+++ src/main/java/com/example/apiweb/Neo4J/repositoryNeo/ICursoNeoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.apiweb.Neo4J.modelNeo.CursoNeoModel;
 
 @Repository
-public interface ICursoNeoRepository extends Neo4jRepository<CursoNeoModel, Long> {
+public interface ICursoNeoRepository extends Neo4jRepository<CursoNeoModel, Integer> {
     // Define custom query methods if needed
 }
